Extract getTableName to separate class
